UPnP NAT traversal, bug fixes, and protocol hardening
- Add UPnP port mapping via igd-next for home router NAT traversal
- PeerManager advertises external address in keepalive messages
- Automatic lease renewal (30 min) and clean shutdown removal
- CGNAT/double-NAT detection with graceful fallback
- Add --disable-upnp CLI flag and BURST_DISABLE_UPNP env var
- Disabled on dev network (matches Nano behavior)
- Assorted bug fixes and protocol implementation improvements

+++ store_lmdb/src/integrity.rs
@@ -67,13 +67,12 @@ pub fn check_integrity(env: &Arc<Env>) -> Result<IntegrityReport, LmdbError> {
                 report.databases_checked += 1;
                 match db.len(&rtxn) {
                     Ok(count) => {
-                        report.total_entries += count as u64;
+                        report.total_entries += count;
                     }
                     Err(e) => {
-                        report.errors.push(format!(
-                            "failed to read database '{}': {}",
-                            db_name, e
-                        ));
+                        report
+                            .errors
+                            .push(format!("failed to read database '{}': {}", db_name, e));
                     }
                 }
             }
@@ -81,10 +80,9 @@ pub fn check_integrity(env: &Arc<Env>) -> Result<IntegrityReport, LmdbError> {
                 // Database doesn't exist yet — acceptable for a fresh node
             }
             Err(e) => {
-                report.errors.push(format!(
-                    "failed to open database '{}': {}",
-                    db_name, e
-                ));
+                report
+                    .errors
+                    .push(format!("failed to open database '{}': {}", db_name, e));
             }
         }
     }
